Removing dependency on rust-bitcoin
There were three situations where rust-bitcoin crate was used in this project:
1. Use of bitcoin hashes library
2. Use of Secp256k1 library
3. Use of bitcoin consensus encoding

In order to shorten compilation time and removes conflicts with downstream crates using different versions of rust-bitcoin it is decided to get rid from this dependency.

The first two are replaced with direct dependencies on the underlying crates. As for the use of bitcoin consensus encoding, it is really specific for bitcoin network protocol and should be discouraged to use in internet2 projects where strict encoding should be used instead (or, at least, lightning encoding, where it is required by lightning network interoperability). Bitcoin network does not uses encryption so communications between future bitcoin nodes should be made via Bifrost using strict and not bitcoin encoding.