InetAddr: Use onion addresses instead of public keys
Define new Tor enum inner types for the `InetAddr` `PartialSocketAddr`, and
`InetSocketAddr` types. Replace the usage of the `TorPublicKeyV3` with
`OnionAddressV3`. While the two types are interchangeable,
`OnionAddressV3` is the human-readable format. For encoding purposes,
`OnionAddressV3` is converted to `TorPublicKeyV3` in order to save some
space.

Additionally, adds port numbers to the `PartialSocketAddr` and
`InetSocketAddr` tor variants. This allows users to specify a specific
port on which an onion address is listening.